GenAI/fastapi-agentic/requirements.txt
fastapi[standard]
pydantic
weaviate-client
langchain
langchain-openai